Removed CCS Verify Scale - Deprecated
