+++ pkg/common/aws/session.go
@@ -1,67 +1,42 @@
 package aws
 
 import (
-	"fmt"
 	"log"
 	"sync"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/aws/aws-sdk-go/service/ec2"
+	"github.com/aws/aws-sdk-go/service/iam"
 	viper "github.com/openshift/osde2e/pkg/common/concurrentviper"
 	"github.com/openshift/osde2e/pkg/common/config"
 )
 
-const (
-	metricsAWSAccessKeyID     = "metrics.awsAccessKeyId"
-	metricsAWSSecretAccessKey = "metrics.awsSecretAccessKey"
-	metricsAWSRegion          = "metrics.awsRegion"
-
-	metricsAWSAccessKeyIDEnv     = "METRICS_AWS_ACCESS_KEY_ID"
-	metricsAWSSecretAccessKeyEnv = "METRICS_AWS_SECRET_ACCESS_KEY"
-	metricsAWSRegionEnv          = "METRICS_AWS_REGION"
-)
-
-type awsSession struct {
+type ccsAwsSession struct {
 	session *session.Session
+	iam     *iam.IAM
+	ec2     *ec2.EC2
 	once    sync.Once
 }
 
-// AWSSession is the global AWS session for interacting with S3.
-var AWSSession awsSession
+// CCSAWSSession is the global AWS session for interacting with AWS.
+var CcsAwsSession ccsAwsSession
 
-func init() {
-	viper.BindEnv(metricsAWSAccessKeyID, metricsAWSAccessKeyIDEnv)
-	config.RegisterSecret(metricsAWSAccessKeyID, "metrics-aws-access-key")
-
-	viper.BindEnv(metricsAWSSecretAccessKey, metricsAWSSecretAccessKeyEnv)
-	config.RegisterSecret(metricsAWSSecretAccessKey, "metrics-aws-secret-access-key")
-
-	viper.BindEnv(metricsAWSRegion, metricsAWSRegionEnv)
-	config.RegisterSecret(metricsAWSRegion, "metrics-aws-region")
-}
-
-func (a *awsSession) getSession() (*session.Session, error) {
+// GetAWSSessions returns a new AWS type with the first AWS account in the config file. The session is cached for the rest of the program.
+func (CcsAwsSession *ccsAwsSession) GetAWSSessions() error {
 	var err error
 
-	// Initialize this once, and initialize it in getSession so that osde2e capabilities that don't use S3
-	// don't have to worry about the noise or configuring S3 properly. The cost here is that things will fail late
-	// on misconfiguration, but since we're the primary consumers of osde2e, at the moment this isn't a big deal.
-	a.once.Do(func() {
-		// We're using static credentials here so that we can use AWS credentials for cluster providers.
-		// When we have more time, we should make this not metrics focused, as the intent of this library is to be purpose agnostic.
-		a.session, err = session.NewSession(aws.NewConfig().
-			WithCredentials(credentials.NewStaticCredentials(viper.GetString(metricsAWSAccessKeyID), viper.GetString(metricsAWSSecretAccessKey), "")).
-			WithRegion(viper.GetString(metricsAWSRegion)))
-
-		if err != nil {
-			log.Printf("error initializing AWS session: %v", err)
-		}
+	CcsAwsSession.once.Do(func() {
+		CcsAwsSession.session, err = session.NewSession(aws.NewConfig().
+			WithCredentials(credentials.NewStaticCredentials(viper.GetString(config.AWSAccessKey), viper.GetString(config.AWSSecretAccessKey), "")).
+			WithRegion(viper.GetString(config.CloudProvider.Region)))
+		CcsAwsSession.iam = iam.New(CcsAwsSession.session)
+		CcsAwsSession.ec2 = ec2.New(CcsAwsSession.session)
 	})
-
-	if a.session == nil {
-		err = fmt.Errorf("unable to initialize AWS session")
+	if err != nil {
+		log.Printf("error initializing AWS session: %v", err)
 	}
 
-	return a.session, err
+	return nil
 }
